refactor(relay): fold the fetch range queries back into track_cache.rs
The split into track_cache/fetch.rs existed only to satisfy the file
length limit, which is gone; the two methods are ledger read queries and
read best next to the other TrackCache API.
